feat: enhance React app with improved configuration management and new components
- Introduced centralized AppConfig class for managing application settings and environment variables.
- Added DialogHandler for handling notifications and confirmations using Ant Design components.
- Refactored RouterLoader to support dynamic route configurations and lazy loading of components.
- Updated various API adapters to utilize the new AppConfig structure for better maintainability.
- Removed outdated DialogHandler implementation and reorganized related files for improved clarity.
