--- packages/create-app/templates/react-app/src/base/adapters/DialogHandler.ts
@@ -1,1 +1,0 @@
-
