+++ packages/create-app/templates/react-app/src/base/apis/userApi/UserApiAdapter.ts
@@ -1,11 +1,11 @@
+import type { AppConfig } from '@/base/cases/AppConfig';
 import { IOCIdentifier } from '@/core/IOC';
 import { RequestAdapterFetch } from '@qlover/fe-corekit';
 import { inject, injectable } from 'inversify';
-import { AppConfigImpl } from '@/core/AppConfig';
 
 @injectable()
 export class UserApiAdapter extends RequestAdapterFetch {
-  constructor(@inject(IOCIdentifier.AppConfig) appConfig: AppConfigImpl) {
+  constructor(@inject(IOCIdentifier.AppConfig) appConfig: AppConfig) {
     super({
       responseType: 'json',
       baseURL: appConfig.feApiBaseUrl
