+++ packages/create-app/templates/react-app/src/base/cases/AppError.ts
@@ -1,8 +1,5 @@
 import { ExecutorError } from '@qlover/fe-corekit';
 
-/**
- * 统一当前项目中开发的错误
- */
 export class AppError extends ExecutorError {
   constructor(
     public readonly id: string,
