+++ packages/create-app/templates/react-app/src/core/globals.ts
@@ -1,9 +1,13 @@
 // ! global variables, don't import any dependencies and don't have side effects
 import { JSONStorage, JSONSerializer, SyncStorage } from '@qlover/fe-corekit';
 import { ColorFormatter, ConsoleHandler, Logger } from '@qlover/corekit-bridge';
-import { DialogHandler } from '@/base/bridge/DialogHandler';
+import { DialogHandler } from '@/base/cases/DialogHandler';
+import { loggerStyles } from '@config/common';
+import { AppConfig } from '@/base/cases/AppConfig';
 
 const isProduction = import.meta.env.VITE_USER_NODE_ENV === 'production';
+
+export const appConfig = new AppConfig();
 
 export const dialogHandler = new DialogHandler();
 
@@ -11,7 +15,7 @@ export const dialogHandler = new DialogHandler();
  * Global logger
  */
 export const logger = new Logger({
-  handlers: new ConsoleHandler(new ColorFormatter()),
+  handlers: new ConsoleHandler(new ColorFormatter(loggerStyles)),
   silent: isProduction
 });
