+++ packages/create-app/templates/react-app/src/uikit/components/RouterRenderComponent.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from 'react';
 import { Loading } from './Loading';
 import BaseRouteProvider from '@/uikit/providers/BaseRouteProvider';
-import { RouterLoaderRender } from '@/base/cases/router-loader';
+import { RouterLoaderRender } from '@/base/cases/RouterLoader';
 
 export const RouterRenderComponent: RouterLoaderRender = (route) => {
   const Component = route.element();
