+++ packages/create-app/templates/react-app/src/uikit/controllers/UserController.ts
@@ -9,7 +9,7 @@ import { inject, injectable } from 'inversify';
 import { IOCIdentifier } from '@/core/IOC';
 import { LoginInterface } from '@/base/port/LoginInterface';
 import { UserApi } from '@/base/apis/userApi/UserApi';
-import { AppError } from '@/base/cases/appError/AppError';
+import { AppError } from '@/base/cases/AppError';
 import { LOCAL_NO_USER_TOKEN } from '@config/Identifier.Error';
 import { SliceStore } from '@qlover/slice-store-react';
